requirements.txt
APScheduler==3.11.2
Authlib==1.6.9
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
Flask-Bcrypt==1.0.1
Flask-Limiter==4.1.1
Flask-WTF==1.2.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
MarkupSafe==3.0.3
ordered-set==4.1.0
packaging==26.0
psycopg2-binary==2.9.11
pycparser==3.0
python-dotenv==1.2.2
requests==2.32.5
resend==2.23.0
soupsieve==2.8.3
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.6
wrapt==2.1.2
WTForms==3.2.1
